refactor: create repositorytest package
